A discord bot that can play a guess the country game
This bot can engage it's players by presenting them with images with geographic locations and it challenges them to guess the country that is shown.
The commands that the user can give the bot are:
/ping : the bot returns "Pong!". This command is used just to test if the bot is online and ready to play
/start [number of rounds] : this starts the game with the number of rounds the user wants to play. It will fetch the images and the game will start. At this moment the timer starts as well and it will be displayed on the screen for each round (2 minutes for each)
/guess [country name] : the player can give the guess command and the country that they are guessing and the bot will say if it's the wrong or correct answer. If the answer is correct, the game will continue to the next round
When the rounds end, there will be a leader board displayed with all the players and their scores.